refactor: History now is a value object
