+++ backend/src/main/java/com/tkpm/sms/application/service/implementation/EnrollmentServiceImpl.java
@@ -20,6 +20,7 @@ import com.tkpm.sms.domain.repository.EnrollmentRepository;
 import com.tkpm.sms.domain.repository.TranscriptRepository;
 import com.tkpm.sms.domain.service.validators.CourseDomainValidator;
 import com.tkpm.sms.domain.service.validators.EnrollmentDomainValidator;
+import com.tkpm.sms.domain.valueobject.History;
 import com.tkpm.sms.domain.valueobject.Score;
 import jakarta.transaction.Transactional;
 import lombok.AccessLevel;
