+++ backend/src/main/java/com/tkpm/sms/domain/model/Enrollment.java
@@ -1,5 +1,6 @@
 package com.tkpm.sms.domain.model;
 
+import com.tkpm.sms.domain.valueobject.History;
 import com.tkpm.sms.domain.valueobject.Score;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
@@ -14,5 +15,6 @@ public class Enrollment {
     Integer id;
     Student student;
     Course course;
+    History history;
     Score score;
 }
