+++ backend/src/main/java/com/tkpm/sms/infrastructure/persistence/mapper/HistoryPersistenceMapper.java
@@ -1,6 +1,6 @@
 package com.tkpm.sms.infrastructure.persistence.mapper;
 
-import com.tkpm.sms.domain.model.History;
+import com.tkpm.sms.domain.valueobject.History;
 import com.tkpm.sms.infrastructure.persistence.entity.HistoryEntity;
 import org.mapstruct.Mapper;
