+++ backend/src/main/java/com/tkpm/sms/infrastructure/persistence/repository/EnrollmentRepositoryImpl.java
@@ -3,10 +3,10 @@ package com.tkpm.sms.infrastructure.persistence.repository;
 import com.tkpm.sms.domain.common.PageRequest;
 import com.tkpm.sms.domain.common.PageResponse;
 import com.tkpm.sms.domain.model.Enrollment;
-import com.tkpm.sms.domain.model.History;
 import com.tkpm.sms.domain.repository.EnrollmentRepository;
 import com.tkpm.sms.domain.repository.SettingRepository;
 import com.tkpm.sms.domain.repository.SubjectRepository;
+import com.tkpm.sms.domain.valueobject.History;
 import com.tkpm.sms.infrastructure.persistence.entity.HistoryEntity;
 import com.tkpm.sms.infrastructure.persistence.jpa.EnrollmentJpaRepository;
 import com.tkpm.sms.infrastructure.persistence.jpa.HistoryJpaRepository;
